refactor: :building_construction: Split modules
Split command and lib modules into more specific modules, help with TS binding

Breaks front-end, go to next commit
