+++ src-tauri/src/command.rs
@@ -1,47 +1,23 @@
 //! All things related to Rust of the Super Mouse AI app
 
 // Crate level use (imports)
-use crate::mutter::{Model, ModelError};
-use log::error;
-use mouce::{
-    common::{MouseButton, MouseEvent},
-    Mouse, MouseActions,
+use crate::{
+    events::MouseClickEvent,
+    mutter::ModelError,
+    types::{AppState, MouseButtonType},
 };
+use log::error;
+use mouce::{common::MouseEvent, Mouse, MouseActions};
 use rodio::{Decoder, OutputStream, Sink};
-use serde::{Deserialize, Serialize};
-use std::{collections::HashMap, fs::File, io::BufReader, path::PathBuf};
-use tauri::{AppHandle, Emitter, State};
-
-/// "Global" state for the application.
-///
-/// This holds all data that is expected to
-/// persist throughout the app's runtime.
-pub struct AppState {
-    model: Model,
-    sound_map: HashMap<String, PathBuf>,
-}
-
-impl AppState {
-    pub fn new(model: Model, sound_map: HashMap<String, PathBuf>) -> Self {
-        // Load model into memory by evaluating short silence
-        // FIXME: Need to do this in another thread, otherwise UI freezes
-        // let _ = model.transcribe_pcm_s16le(&[0.0; 20_000], false, false, None, None, None);
-        AppState { model, sound_map }
-    }
-
-    /// Get sound by the provided name or by prepending `default_` to the beginning.
-    pub fn get_sound_path(&self, sound_name: &str) -> Option<&PathBuf> {
-        self.sound_map
-            .get(sound_name)
-            .or_else(|| self.sound_map.get(&format!("default_{}", &sound_name)))
-    }
-}
+use std::{fs::File, io::BufReader};
+use tauri::{AppHandle, State};
+use tauri_specta::Event;
 
 #[tauri::command]
 #[specta::specta]
 /// Take WAV audio data and transcribe it with application Whisper model.
 ///
-/// Check [`crate::mutter::Model`] for details on argument
+/// Check [crate::mutter::Model::transcribe_audio] for details on arguments
 pub async fn transcribe(
     app_state: State<'_, AppState>,
     audio_data: Vec<u8>,
@@ -115,27 +91,6 @@ pub async fn play_sound(app_state: State<'_, AppState>, sound_name: String) -> R
     Ok(())
 }
 
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-/// Enum representing mouse button type
-///
-/// Main three are left, middle, and right
-#[non_exhaustive]
-enum MouseButtonType {
-    Left,
-    Middle,
-    Right,
-}
-
-impl From<&MouseButton> for MouseButtonType {
-    fn from(mb: &MouseButton) -> Self {
-        match mb {
-            MouseButton::Left => MouseButtonType::Left,
-            MouseButton::Middle => MouseButtonType::Middle,
-            MouseButton::Right => MouseButtonType::Right,
-        }
-    }
-}
-
 /// Function to listen for any clicks from mouse and emits Tauri event.
 ///
 /// Returns either the callback id or an error message.
@@ -144,8 +99,9 @@ impl From<&MouseButton> for MouseButtonType {
 pub fn listen_for_mouse_click(app_handle: AppHandle) -> Result<u8, String> {
     Mouse::new()
         .hook(Box::new(move |e| match e {
-            MouseEvent::Press(button) => app_handle
-                .emit("mouse_press", MouseButtonType::from(button))
+            MouseEvent::Press(button) => 
+            MouseClickEvent::with_payload(MouseButtonType::from(button))
+                .emit(&app_handle)
                 .map_err(|e| {
                     error!("App Handle expected to emit press event with button playload but could not: {}", e);
                 })
@@ -154,49 +110,4 @@ pub fn listen_for_mouse_click(app_handle: AppHandle) -> Result<u8, String> {
             _ => (),
         }))
         .map_err(|err| err.to_string())
-}
-
-/// Given a key, check if it matches one of the (specific) modifier keys.
-///
-/// The main modifiers are: Alt, Control, Meta, Option, and Shift (both left and right).
-pub fn is_modkey(key: &device_query::Keycode) -> bool {
-    use device_query::Keycode as K;
-    matches!(
-        key,
-        K::Command
-            | K::LAlt
-            | K::LControl
-            | K::LMeta
-            | K::LOption
-            | K::LShift
-            | K::RAlt
-            | K::RControl
-            | K::RMeta
-            | K::ROption
-            | K::RShift
-    )
-}
-
-#[derive(Clone, Debug, Default, Serialize)]
-/// Information about modifier key event (pressed or released)
-pub struct ModKeyEvent {
-    key: String,
-    is_pressed: bool,
-}
-
-impl ModKeyEvent {
-    /// New Modifier Key Event
-    pub fn new(key: String, is_pressed: bool) -> Self {
-        Self { key, is_pressed }
-    }
-
-    /// New pressed event for given key
-    pub fn pressed(key: String) -> Self {
-        Self::new(key, true)
-    }
-
-    /// New release event for given key
-    pub fn released(key: String) -> Self {
-        Self::new(key, false)
-    }
 }
